Fix issue with class not being able to cast unless string
I found this on stack overflow https://stackoverflow.com/questions/47255419/rails-inverse-of-polymorphic-association-in-concern-gives-typeerror-cant-cast
which suggests that casting classes when prepared_statements are set
to true was broken in rails 5.1.

This means that when you filter by type, you need to pass the class
name as a string as that's how it's stored in the db.