Allow models to determine their own public route
We are currently using `url_maker` to determine the public route of a
given document.

This code is complicated and relies on route helpers being defined
either in the routes file, or elsewhere within an included helper.

By allowing models to define their own public routes, we can move away
from a system of relying on route helpers and eventually retire
`url_maker`.