Fix broken test
This test's setup block has `I18n.with_locale("de")` however the test
fails to append the `de` locale to the URL when checking the base path.

It appears this was because the relevant presenter also never included
the locale when using `UrlMaker`.  Now we are switching away from
`UrlMaker`, this error became apparent.