Always use the GOV.UK public URL
Whitehall's `UrlMaker` allows us to define a custom public website
domain and protocol.  This is useless as all public content is served
from https://www.gov.uk (or the equivalent for the other environments)
and we are using Plek to always give us the correct root.

Therefore updating all tests to use only the root given by Plek.