Return 404 on detailed guide routes
This relates to the problem addressed in: https://github.com/alphagov/whitehall/pull/7113

If Whitehall receives a request for a detailed guide it will log an
error of:

```
uninitialized constant DetailedGuidesController

      Object.const_get(camel_cased_word)
            ^^^^^^^^^^
Did you mean?  DetailedGuidePresenter

            raise MissingController.new(error.message, error.name)
            ^^^^^
```

Before returning a 404.

This changes that to just return a 404 as a rack response and resolve
wiring this to a non-existent controller. This stops the logging but
continues returning a 404 response.