Add controller actions and view for reordering RoleAppointments
This adds the endpoints for reordering a persons role appointments to
the persons controller.

These endpoints should only be accessible by users with the 'GDS Editor'
permission.

Only active/current role appointment should be able to be reoredered.

I've got with a semi-crude implementation, but it matches the
current strategy as much a possible while not using the id.

It simply grabs all the ids from the current role appointments, and
assigns the order of each appointment based on the ordering passed
thorough to the controller.