Update to use EditionSlugController
While looking through the previous PR to update the dev docs,
it occured to me that as of now were not linking to this page via the UI.

It takes a :document_id in the URL, which GDS Content Designers will
not have access to without developer intervention.

It makes more sense for it to take an edition id instead.

This commit adds an EditionSlugController, which contains the edit_slug
and update_slug actions and wires in the view.

I've also reverted the document_rules and am just using the edition_rules
instead.