Updates govspeak class to meet BEM standards
Previously the class block was `govspeak_help...`, this has now been changed to `govspeak-help...`.

The file was also placed in a components folder within admin, this has now been moved to views. This was incorrectly placed as the govspeak helper is not a component but is a partial that is being used by views.

There was also a file in views folder which was called `govspeak-help` but this was not referenced anywhere and was orphaned. The content of this was replaced with the content of the `components/govspeak-help` file.