Update CI GHA workflow to use reusable components
This updates the CI workflow to use the individual reusable workflows
from govuk-infrastructure, instead of using inputs to skip steps. This
allows workflows to be less noisey with steps that aren't relevant for a
particular app. Also allows more flexiblity in configuration, as extra
jobs can easily be added directly without modifying a shared workflow.