Ensure editing is passed as a query string to the AuditTrailController
This ensures that `editing=true` is passed to the query string of the
pagination link if it was passed as an argument when the view
component was instantiated.

This ensures that the relevant guidance about saving before adding a
remark is rendered on the edit edition and edit translation pages
when the tab is updated.