Add DocumentHistory module for async pagination
This adds async pagination of the history tab and applies it to the
document history tab.

When the older or newer link is clicked, it sends an Ajax request to the
EditionAuditTrailController#index action and updates the document
history tab.

A couple of things to note:

1. We're going to begin retiring support for IE 11. As this module
uses promises which aren't supported we shouldn't override the
default behaviour of pagination and allow the page to reload.

2. If the promise fails, we catch the error and reload the page
with the href in the link that was clicked (older or newer).