Order an organisations promotional features by the ordering attribute
This orders the organisations 'has_many :promotional_features'
association by the ordering value by default.

This means that wherever used, it will show the promotional features
in the expected order to the user, and will pass changes downstream
in the correct order without having to explicitly order throughout the
codebase.