Present Corporate Information Page links as part of Worldwide Organisation content item
In order to migrate the rendering of Worldwide Organisations out of
Whitehall, we need to add further information into the content item.

This adds links to the organisation's Corporate Information Pages as
links, then the display title and order into the details hash.

We cannot rely solely on the links as we were previously sometime
overriding the document's title with a custom string, which is why I've
needed to include that text in the details hash.

There is so much code in the presenter as I've maintained the order they
were previously being presented when rendered by Whitehall.