Add radios for selecting image or youtube vid for promotional_feature_items
This adds radio buttons for selecting whether you want to add an
image or link to a youtube video for a promotional feature item and
handles cleaning the params in the promotional features controller.

Later i'll add functionality to conditionally reveal the image &
youtube video url fields when the corresponding radio is checked.