Ensure image or url is present on promotional_feature_items on save
At the moment, we don't validate that an image is present for a
promotional feature item on save. While the objet itself will
save, when it tries to publish downstream to the PublishingApi
it returns a 500 as the organisation schema expects an image.

As a promotional feature requires either an image, or youtube url,
i've added a validation which checks that one of them is present
on save.