Validate youtube_video_url
This checks thats the youtube_video_url is in a valid format before
we persist it to the database.

It follows the logic laid out in the youtube_id gem.
https://github.com/reu/youtube_id/blob/master/lib/youtube_id.rb

We may decide we want to update this as a dependency is we find a
need for reuse of this functionality across applications.