Include locale for document base path in presented content
When migrating away from using polymorphic paths, the inclusion of the
locale was missed in a call to `public_path`.

This meant we were always returning the English base path, regardless of
the translation.

If the Welsh page was edited, we ended up presenting the Welsh content
on the English translation.  If the English page was edited, we
presented the English content only.  In both cases, the Welsh page was
autoamtically unpublished.