Rename legacy `public_path` method
The `StatisticsAnnouncement` model contains a `public_path` method that
is not the same as the new `public_path` method where models know their
own routes.

It is in fact the same as the `base_path` method on other models.

Therefore renaming it so we don't call the wrong method when determining
public paths.

This was only noticed after `PolymorphicPath` started including the
locale, since there was then an unexpected number of arguments.