Run verify-pact workflow only on push to main
This updates the workflow to not run on every commit pushed, instead
only on pushes to main (i.e. merge a pull request). This is rarely
needed and is a waste of compute resources. It also creates extra noise
on pull request status checks, as it generates two runs when a PR is
created.