Update translated path test for World Location News
During a recent incident, we saw either the Welsh translation of an
organisation page appear on the English content item and/or the Welsh
page unpublished and redirected to the English page.

After the incident we added tests for organisation pages to ensure the
correct base path and routes were being presented to Publishing API for
translated organisations.

There was already a test for the base path of World Location News, but
not for the routes hash.  This updates that test to be consistent with
the other presenter tests.