Add translated path test for World Locations
During a recent incident, we saw either the Welsh translation of an
organisation page appear on the English content item and/or the Welsh
page unpublished and redirected to the English page.

After the incident we added tests for organisation pages to ensure the
correct base path and routes were being presented to Publishing API for
translated organisations.

This adds the same tests for World Locations.