Allow emails to be intercepted
We don't want emails sent in non-production environments to be delivered
to end users. For example, email notifications sent in integration and
staging should not be delivered in case they're mistaken as production
notifications.

Therefore, the GOV.UK Notify accounts used by Whitehall in integration
and staging have been configured to reject emails being sent to
recipients who aren't on a pre-defined 'allow list' of addresses.

This commit adds a Mail Recipient Interceptor, which will intercept
emails before Rails attempts to deliver them and update the recipient.
This can be used in integration and staging by setting the environment
variable `EMAIL_ADDRESS_OVERRIDE` to an email address that is known to
be on the GOV.UK Notify 'allow list'.

This setup was _heavily inspired_ by the Content Publisher app, which
also implements an interceptor:

- [lib/mail_recipient_interceptor.rb][1]
- [config/initializers/mail_recipient_interceptor.rb][2]

[1]: https://github.com/alphagov/content-publisher/blob/efc7cc8156d153d458472e5dd1f81992241c0b87/lib/mail_recipient_interceptor.rb
[2]: https://github.com/alphagov/content-publisher/blob/efc7cc8156d153d458472e5dd1f81992241c0b87/config/initializers/mail_recipient_interceptor.rb