Clean up @design-system-wip feature specs
This goes through the feature specs, and removes the @design-system-wip
tag and fixes the tests where possible.

The following tests only required the tag to be removed and passed:

1. Edition attachments feature spec
2. Force publishing editions feature spec
3. New tagging workflow feature spec
4. Policy groups feature spec

The following tests required small changes to fix once the tag was
removed:

1. Admin filtering documents feature spec
2. Consultation feature specs
3. Admin statistical data sets feature spec
4. Admin world locations feature spec
5. Document collections feature spec
6. Fatalities feature spec
7. News article feature spec
8. Ppeech feature spec

There's still some JS that needs to be added for speeches before
we can remove the final @design-system-wip from the tests.