Add 'setupSpeechSubtypeEventListeners' to the edition-form module
This adds event listeners to check that when authored_article is
selected as the speech subtype the location field is reset and hidden,
and the content on the page which references is speaker is set to
Writer.

It also handles the select being changed from authored_article to
another subtype. In this case the location field becomes visible and
the content referencing writer is updated to speaker.