Add presenter for Historial Accounts
This will be used to put the content of Historial Accounts (i.e. profile
pages for past prime ministers) into Publishing API to enable us to
serve the content from an application outside of Whitehall.

The rendering app will remain as Whitehall until the frontend has been
built for this page.