Stop checking Govspeak HTML output for safety
The SafeHtmlValidator checks values for unsafe HTML.

For HTML attachments, this validator has been running both on the
Govspeak input as well as the generated HTML output.

However, it's possible (and perfectly valid) for generated HTML output
to contain HTML that would be considered unsafe if entered by a user.
When Govspeak is converted to HTML, it's possible for
govuk_publishing_components to be included – for example, when adding
a [button] to a document – or indeed any other HTML that might be added
in post-processing.

This commit changes the validation rule so it only applies to Govspeak
input provided by the user. Govspeak HTML output is no longer validated.
This is safe to do because Govspeak post-processing is trusted code. For
more information, see alphagov/govspeak#237 which established this as
safe behaviour.

[button]: https://github.com/alphagov/govspeak#button