Fix bug with edit edition history
At the moment, the legacy edit page doesn't set the history variables
in the controller correctly as it's being passed the new merged history
object.

We should only be passing the new merged history for the page when
to the edit page when the "Preview design system" permission is present
on the user.