Ensure document history pagination works with 'Preview next release' flag
Currently, the EditionAuditTrailController#index action is hit under 3
from 3 different pages.

1. edit edition
2. edit edition translation
3. edition summary

These pages are going out in the following releases

Release 1.5

1. edition summary

Release 1.6

1. edition edition
2. edit edition translation

Due to these pages going out in different releases, we need a mechanism
to flag the new behaviour to the edition summary page, but not the two
edit pages.

Fortunately the editing="true" query string param is passed through to
the endpoint for the two edit pages so we can use that to ensure
the old flow is retained until the new pages go out in 1.6.