Populate images index page
This commit creates some placeholder content for
the images index endpoint. This is mostly non-functional
for the time being, function will be built in at a later
stage.

The commit also adds some logic to the controller which
provides the Design System layout, checks the user's
permissions and manipulates the edition to have an
empty image if there are no images attached.