Show World Taxons on the Edition Summary page
For some reasons the World Taxons are all have
`@visible_to_departmental_editors=false`.

This is causing then not to show the Edition Summary page. As far as
we can tell there is no reason for this.

This removes checking that this is set to true before rendering them
on the summary page.