Update link anchor to open 'Fact checking' tab
The fact check response email contains a link that publishers can click
to see the comment left by the reviewer.

However the link anchor has changed following a move to the GOV.UK
Design System, which resulted in element IDs changing on the page.

This new link will automatically show and jump to the 'Fact checking'
tab so users can see the comment.