Ensure new broken links report cannot be generated from edit page
At the moment, if a broken links report is pending when the edit page
is loaded, the JS module is loaded and starts polling the
LinkCheckReportsController#show endpoint.

Once the report is completed it inserts the '_link_check_report' partial
into the innerHTML of the link reports div.

This causes the button to appear on the edit page when it shouldn't be.

To work around this we can check if the requests referrer is the edit
edition page. If this is the case we can set the allow_new_report locale
to false to ensure the new report button doesn't render.