Add a filter to the Document History tab
This adds a dropdown/select so users can filter the types of entries
they see in the Document History tab. Where previously this tab merged
History and Internal Notes, it's now possible for users to choose from
seeing:

- Everything (i.e. History and Internal Notes together)
- Document history only
- Internal notes only

This filter has been progressively enhanced, so that it's automatically
applied when the <select> has been changed if the user has JavaScript.
Without it, the user will need to click the "Filter" button.

How it works
============

When a user selects an option, a request is made to the
EditionDocumentHistoryController#index endpoint with the 'only'
value as a param in the query string.

This is passed into the PaginatedTimeline object as the 'only'
parameter.

Without JavaScript, or if the Ajax call fails, the page is reloaded and
the the filter param is passed into the timeline object in the
'fetch_version_and_remark_trails' method.

Changes to admin-history-pagination feature spec
================================================

Test coverage has been added to cover the filter journey, both for
non-JS and JavaScript users. It combines filtering with pagination to
ensure the filter is persisted between pages.

I've also taken the opportunity to refactor and tidy up the feature test
a little. The intention was to make the admin-history-pagination steps
simpler to follow and extract useful functionality into reusable helper
method.