Include past pm role appointments without historical accounts in the content item
Since historical accounts are published as separate pages, to include information on past prime ministers with historical accounts, we can just link to these content items and rely on link expansion in publishing api to ensure the relevant information is displayed on the page.

However, a historical account may not be present for more recent past prime ministers, since it relies on the content being created. We want to ensure these past prime ministers still display on the page, but we cannot rely on linking to an existing page for these. Instead, we present the relevant information for these prime ministers in the details hash. Associated changes have been made to the content schema in publishing api to allow for this.