Do not raise `GdsApi::HTTPUnprocessableEntity` for Worldwide Organisations
We know there is an issue with Worldwide Organisations having the same
base path as one of their Corporate Information Pages. This leads to
`GdsApi::HTTPUnprocessableEntity` being raised when the Worldwide
Organisation itself is published again.

There is currently no user facing impact as the content item is not
being used for anything on these pages.

This is tracked on a Trello card looked after by Publishing Platform
(see https://trello.com/c/qYYHTFW0). When rendering is moved out of
Whitehall, this issue will need to be fixed.

However, in the meantime, we are sending approximately 46 events each
day to Sentry. This both uses our allowance of events and is a nuisance
in Slack.

This will stop an error being raised solely for the known scenario
covered by the aforementioned Trello card, which should reduce some of
the noise in Slack.