Only present past dates of relevant role in historical accounts index presenter
On the past prime ministers page, we display their dates in role.

For past prime ministers without historic appointments, in the content item we were presenting their dates in role for any historic role they held (for example, dates when they were chancellors). This fixes that.

Note that since we're not yet using this content item, this isn't affecting the data displaying on the website, which is still accurate.