Ensure Autoselect passes blank value to controller if none selected
At the moment, when multiple is set to true on the autocomplete a
blank value is not passed to the controller.

This means that it's not possible to set the fields from a value to nil
on the object.

There's 2 ways to handle this

1. pass a blank hidden field which is overwritten when a value is
present on the field
2. construct the blank params in the controller

Passing a hidden field is the component itself seems like the
easiest and robust solution for extensibility.