Ensure flash-danger Google Analytics event is retained
We've moved away from using flash[:alert] in favour of the error
summary component when there are errors on an object.

However, we want to retain the GA event that is triggered by the flash.
To get around this I've ensured a flash["alert"] is only rendered on
the page when the error summary block is empty.

This will ensure we only ever get 1 banner on the page. To ensure we
still get the GA event i've wrapped the error summary component in a
div which tracks the flash["alert"].