Add #login_as_preview_design_system_user helper method
This adds a helper method which creates a user with the
User::Permissions::PREVIEW_DESIGN_SYSTEM permission
based on the role passed in as a paramter.

For example if you pass :managing_editor in. It will create a user
from the managing_editor factory with the permission and log the user
in.