Make locale a keyword argument for organisations and worldwide organisations
Anywhere where we were not passing options through when we were previously calling public_path, we are safe to just pass the default locale (i.e. english) through.
In the places where we need to call organisations with translations, the calling code has been slightly altered to ensure that other arguments previously suplied in an options hash are separated into an explicit hash from the locale, which will now be interpreted as a keyword argument