Make locale mandatory for editions
This is being done as part of an incident action, to ensure that we always specify locales when we create paths for documents.

This involved some changes in the public document routes helper. Here, the locale is not made mandatory, and is instead introduced as a keyword argument with a default. This is to pull the locale out of the options hash, but have the ability not to specify it since we don't want to override the existing locale selecting logic in here.