Add document collections to groups
Document collections are whitehall documents. However there is a validation
on the document collection model that prevents document collections being added
to a document collection group [1].

However there are lots of document collections of document collections live on gov.uk.
This is achieved via a workaround in the UI: A user adding a document collection URL to
the "Add document to a group" form, sees the error "Document cannot be a document collection."
But adding a document collection url to the field "Add GOV.UK content created outside of Whitehall to a group"
creates a DocumentCollectionNonWhitehallLink and the document is able to be added.

This rake task follows that same pattern.

[1] https://github.com/alphagov/whitehall/blob/main/app/models/document_collection_group_membership.rb#L2