Move republishing of How Government Works page
At the moment, this page gets republished when the other static pages
are published.

However we plan to start including dynamic content in this page's
content item.

Therefore moving the republishing so it can be called when required
(e.g. when an organisation is added or removed) and republished
on-demand using a rake task.

This is similar to the Past Prime Ministers index page.