Update How Government Works page when ministerial roles change
The page includes a count of the number of ministerial roles appointed,
so we need to republish the page when any ministerial appointments are
changed.

The `set_order` method has been changed to skip callbacks (by using
`update_column` instead of `update`) as this caused the How Government
Works page to be republished twice.