Add roll call introduction to fatality notice presenter
While this is not rendered on the fatality notice page, we will need it present in the fatality notice content item when rendering the fields of operation pages, which present this information for each fatality notice. Adding it in to the content item here will allow us to link to a fatality notice in the field of operation and add the roll call introduction to the expansion rules.

Additionally, we may choose to render this information on the fatality notice page in future, so having it in the content item makes sense.