Fix bug with alternativeFormatProviderId passing "null"
At the moment, we always append `alternative_format_provider_id`
to the query string when the preview endpoint is hit regardless of
whether it's null.

This results in params[:alternative_format_provider_id] always being
true as "null" is passed as the query string.

This only sends it through if a value if present.