Add 'Select with Search' component
This adds an app-specific "govuk_publishing_component" which largely
extends the "Select" component.

It supports the same parameters and functionality as the Select
component, but the generated <select> is enhanced with Choices.js to
make it filterable.

The Choices.js implementation is themed to look like a GOV.UK styled
component, so it fits in with the surrounding page decor.

Selects created with this component also support grouped options. To
provide grouped options, pass in the 'grouped_options' parameter instead
of the usual 'options' parameter.

I've included test coverage of the JavaScript module
"select-with-search" as well as an integration test to cover generated
HTML output.