Implement 'reshuffle mode' for 'How Government Works' page
This page includes counts of the current ministers, departments and the
current PM.

If there is a ministerial reshuffle going on, we do not want this
information showing in the content item as it may be out of date or
include information that has not yet been announced publicly.

Therefore removing all the details and links when reshuffling.