Fix crest on Worldwide Organisation pages
At the moment, we display the simplified royal coat of arms as the crest
on all Worldwide Organisation pages.

This is not correct as some organisation have different logos (e.g.the
Department for Business & Trade).

Updating this so we match the logo that is used by the worldwide
organisation's first supporting organisation.