Add column to the Editions table
We will soon be converting some specialist topics into document collections.
This will be done via a rake task to create a draft document collection.

Storing a reference to the source topic will make it possible to check if a topic
has already been converted, and avoid accidentally creating duplicates.

This is a temporary association, once the conversions have all been completed,
the column can be removed again.