Adapt image cropper for client side cropping
Images can't be cropped server side because Whitehall doesn't store
full-size original images, unlike Content Publisher. Therefore cropping
needs to be handled on the client side.

This adapts the image cropper component to crop images directly in the
user's browser, and then attach the resultant image to a hidden file
input so it gets uploaded to the server with the correct dimensions.

Co-Authored-By: Ollie Treend <ollie.treend@digital.cabinet-office.gov.uk>