Don't add unpublished documents to groups
When a piece of content is tagged to a specialist topic, and added to a specialist topic group,
the collections publisher database will not be updated should that content be unpublished.
This means that the specialist topic content item will still contain links to pages that are no longer live.

Example: https://www.gov.uk/api/content/topic/benefits-credits/tax-credits

details[:groups].first[:content_ids] contains 4 content ID's. Only three are rendered on the
live page. The last content id is a redirect (https://www.gov.uk/api/content/qualify-tax-credits), and as
such it's not indexed in search so will be hidden from the frontend by logic in Collections[1]

We want our document collection to represent the live specialist topic page as seen by users.
We are not concerned with preserving old associations. Therefore we should skip the addition of
any Whitehall documents that have been unpublished. And any non whitehall documents that are redirects.

[1] https://github.com/alphagov/collections/blob/main/app/models/list_set.rb#L49-L59