Ignore 'document type' unless slug is ambiguous
The existing behaviour was slightly unintuitive, because it would match
a document with a unique slug so long as the 'document type' field was
empty. However if the 'document type' field contained an invalid value
(for example, a typo), the match would fail - despite the slug being
unique.

In practice this meant that typos and erronous document types in the
provided CSV would cause rows to fail, even if it would have been
possible to identify the document by slug alone.

I've therefore tweaked this behaviour to make it more resilient, so
matches can still happen when the slug is unique.

I've also taken this opportunity to refactor the code to make the flow
of logic a bit easier to follow.