Add support for retagging Statistics Announcements
Statistics Announcements are not Documents in Whitehall. But they still
have organisations, and still need to be retagged when Machinery of
Government changes happen.

So I've extended the 'Bulk Organisation Updater' to add support for
Statistics Announcement documents.

Note: the use of `rubocop:disable Rails/SaveBang` is a bit of a smell.
It indicates that we don't actually check that the database save was
successful. However, this is in keeping with the existing behaviour of
this class - i.e. it's already used by the `update_edition` method.

So I've decided to use the same approach so as to be consistent with the
existing behaviour. If we find this behaviour to be problematic, then we
should change it in both places.