Skip additional callbacks in database seeds
When running the seeds on a fresh project install (such as through `make
whitehall` in `govuk-docker`), there are several calls made to
`publishing-api` which fail, causing the seeds to break and stop running
any further.

Looking through the seeds, these calls to `publishing-api` are
unintentional and the relevant callbacks should be skipped as they
already are for some of the seeds.