Change setup for GovernmentPresenterTest
This test was previously (incorrectly) inheriting from
`PresenterTestCase`, which is used only for user-facing presenter units
(not Publishing API presenter tests).

As a result, writing to the database was disabled meaning we never set
the `updated_at` timestamp on records. This resulted in schema tests
(which were previously not present) failing as this value was always
nil.

Therefore updating to inherit from `TestCase` and correctly namespace
the test.