Move Government presenter test cases into a context block
There were some tests outside of a context block, meaning they still
existed when the contexts were run. This led to a validation error on
the model as there are overlapping governments.

This was never an issue before as the test was incorrectly set up to not
write anything to the database, meaning the validation never ran.