Add presenter for World index page
This page does not currently have a content item, which is required to
render it outside of Whitehall.

Due to complications with the World Taxonomy (i.e. there is no content
item that contains both the country name and correct base path or slug),
we cannot include the locations as links and have therefore put them
into the details hash.