Remove Jenkinsfile to switch off Jenkins CI
Now that we use GitHub Actions in the new EKS environment, we no longer
need Jenkins CI to run.

If I understand the behaviour of alphagov/govuk-saas-config correctly, I
think we need to remove the Jenkinsfile from this repo so it doesn't
enforce Jenkins CI checks to pass on PRs.