Quick fix to enable branch protection rules
This is a workaround to make govuk-saas-config work correctly with the CI jobs defined in this repo.

govuk-saas-config expects a single job called "test" to exist, which must pass for PRs to be merged. Whereas the #govuk-replatforming team have implemented CI as a series of multiple jobs.

This change bridges the gap between the two, by defining a job called "test" that will only pass once all the other CI jobs have passed.