Revert "Remove Jenkinsfile to switch off Jenkins CI"
It is [not currently recommended](https://gds.slack.com/archives/CD6JUFXML/p1680013649936339?thread_ts=1680003113.239669&cid=CD6JUFXML) that Jenkinsfiles are removed from repositories. So we'll need to keep running CI in both Jenkins and GitHub Actions for now.

This reverts #7484 which removed the Jenkinsfile. That PR was originally intended to unblock the team because Jenkins CI was broken and [failing all builds](https://gds.slack.com/archives/CADKZN519/p1679664795540369).