Add dashboard index view in the GOV.UK Design System
This adds the dashboard index view in the GOV.UK Design System.

As the like for like implementation didn't really work with both
the draft and force published editions taking up half the page
each, i've largely copied the design in Carlos' prototype.

I've fixed the table columns width so that the two tables align.