Update DateTimeFields component to ensure it's clear time is optional
At the moment, we indicate whether a datetime field is required
by adding "(required)" to the fieldset. It's unclear to the user
that time is optional and that only the date fields need to be
completed.

We've decided to resolve this by indicating that date is required,
but time is optional by moving the "(required)" from the fieldset
to the date label.