Update schema for /world/embassies
Previously, this was using a schema_name of 'placeholder'. Placeholder content items are not sent to router, meaning that router did not know about this content item, and when the prefix route for /world was removed, this page 404'd. Updating this to use the barebones schema for embassies to avoid this problem when we republish the /world content item. Later work to migrate this page out of being rendered by whitehall will add additional details to this content item and schema.