Run schema tests against main branch of Publishing API
The concept of the deployed-to-production branch no longer exists in GOV.UK infrastructure since we switched to a Kubernetes platform. Thus this branch is stale and no-longer represents the deployed version of Publishing API.

Switching this to main means that we will be testing against the version of Publishing API that is expected to be deployed - as Publishing API is continuously deployed.