Add "Add another" module
This creates a new component to duplicate fields in a form, based on a
couple of existing features, as shown below.

This is designed to be a generic component though it will initially be
used in this specific location ("New fatality notice" page).
This PR also adds the component to that location.

The main functions it implements are

Adds an "Add" button to the section when the component loads. The actual
text of this button is configurable but defaults to "Add another".
On clicking the "Add" button the existing fields are copied the values
in the copied fields are cleared the values for the id, for and name
attributes are incremented in the newly created fields.

The new fields are appended to the section a "Delete" button is added to
all the fields. The focus is set to the first field of the newly added
set.

On clicking any of the "Delete" buttons the appropriate set of fields is
removed the values for the id, for and name attributes are updated in
all the remaining fields to be sequential.

If only one field is displayed as a result of this action the "Delete"
button no longer displays the focus is set to the first field of the
section.